Update: setup.sh
・すでに存在していたファイルがシンボリックじゃない場合のみバックアップするようにした
・シンボリックリンクがすでにあるときは何もしないようにした

+++ setup.sh
@@ -4,9 +4,14 @@ dotfiles=( .zshrc .tmux.conf .screenrc .vimrc .hgrc .hgignore_global )
 
 for file in ${dotfiles[@]}
 do
-  if [ -e $HOME/$file ]; then
+  if [ -e $HOME/$file -a ! -L $HOME/$file ]; then
+    echo "backup file : ${file}"
     mv $HOME/$file $HOME/$file.bak
   fi
 
-  ln -s $HOME/dotfiles/$file $HOME/$file
+  if [ ! -L $HOME/$file ]; then
+    echo "new symbolic link : ${file}"
+    ln -s $HOME/dotfiles/$file $HOME/$file
+  fi
+
 done
